Restaurant_Reservation/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
dotenv>=0.9.0
drf-spectacular>=0.28.0
psycopg2-binary>=2.9.0